Removed pinned version of docker because of homebrew madness
Whenever we run upgrade, we will fetch the most recent version of
docker and the related tools. We do this because Homebrew decided to
break out their old commits into a legacy-homebrew repo, causing
older versions of docker to become unavailable. Basically we will
always just track the latest version when we do an "upgrade". This is
ok, because I expect native Docker support for OSX to make all this
obsolete.